factor out capture results
