+++ query-server/server/src/main/scala/com/azavea/ca/server/TracksQueries.scala
@@ -27,21 +27,6 @@ object TracksQueries
   val gmTableName = "geomesa.tracks"
   val gmFeatureTypeName = "generated-tracks"
 
-  def capture(isLooseOpt: Option[String], waveOrMesa: String, cql: String): (Option[TestResult], Option[TestResult]) = {
-    val query = ECQL.toFilter(cql)
-    if(waveOrMesa == "wm") {
-      val mesa: TestResult = captureGeoMesaQuery(query, checkIfIsLoose(isLooseOpt))
-      val wave: TestResult = captureGeoWaveQuery(query)
-      (Some(mesa), Some(wave))
-    } else if (waveOrMesa == "w") {
-      val wave: TestResult = captureGeoWaveQuery(query)
-      (None, Some(wave))
-    } else {
-      val mesa: TestResult = captureGeoMesaQuery(query, checkIfIsLoose(isLooseOpt))
-      (Some(mesa), None)
-    }
-  }
-
   def queryRoute(queryName: String, cql: String) = get {
     pathEndOrSingleSlash {
       parameters('test ?, 'loose ?, 'waveOrMesa ? "wm") { (isTestOpt, isLooseOpt, waveOrMesa) =>
